refactored document version

+++ src/main/java/com/example/SVC/controller/DocumentController.java
@@ -10,6 +10,8 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
+import java.math.BigDecimal;
+
 @RestController
 @RequestMapping("/api/documents")
 @RequiredArgsConstructor
@@ -52,7 +54,7 @@ public class DocumentController {
     @PostMapping("/{id}/rollback")
     public ResponseEntity<String> rollbackDocument(
             @PathVariable("id") Long documentId,
-            @RequestParam("version") double versionNumber) {
+            @RequestParam("version") BigDecimal versionNumber) {
 
         try {
             documentService.rollbackDocument(documentId, versionNumber);
